factor more things out of __init__.py
