+++ relengapi/blueprints/tokenauth/__init__.py
@@ -3,8 +3,6 @@
 # file, You can obtain one at http://mozilla.org/MPL/2.0/.
 
 import logging
-import sqlalchemy as sa
-import wsme.types
 
 from flask import Blueprint
 from flask import g
@@ -13,11 +11,13 @@ from flask.ext.login import current_user
 from flask.ext.login import login_required
 from relengapi import apimethod
 from relengapi import p
+from relengapi.blueprints.tokenauth import loader
+from relengapi.blueprints.tokenauth import tables
 from relengapi.blueprints.tokenauth import tokenstr
+from relengapi.blueprints.tokenauth import types
 from relengapi.lib import angular
 from relengapi.lib import api
 from relengapi.lib import auth
-from relengapi.lib import db
 from relengapi.lib import permissions
 from werkzeug.exceptions import BadRequest
 from werkzeug.exceptions import NotFound
@@ -40,70 +40,6 @@ bp.root_widget_template(
     'tokenauth_root_widget.html', priority=100, condition=permitted)
 
 
-class Token(db.declarative_base('relengapi')):
-    __tablename__ = 'auth_tokens'
-
-    def __init__(self, permissions=None, **kwargs):
-        kwargs['_permissions'] = ','.join((str(a) for a in permissions or []))
-        super(Token, self).__init__(**kwargs)
-
-    id = sa.Column(sa.Integer, primary_key=True)
-    description = sa.Column(sa.Text, nullable=False)
-    _permissions = sa.Column(sa.Text, nullable=False)
-
-    def to_jsontoken(self):
-        return JsonToken(id=self.id, description=self.description,
-                         permissions=[str(a) for a in self.permissions])
-
-    @property
-    def permissions(self):
-        token_permissions = [p.get(permissionstr)
-                             for permissionstr in self._permissions.split(',')]
-        # silently ignore any nonexistent permissions; this allows us to remove unused
-        # permissions without causing tokens permitting those permissions to fail
-        # completely
-        return [a for a in token_permissions if a]
-
-
-class TokenUser(auth.BaseUser):
-
-    type = 'token'
-
-    def __init__(self, token_id, permissions):
-        self.token_id = token_id
-        self._permissions = set(permissions)
-
-    def get_id(self):
-        return 'token:#%s' % self.token_id
-
-    def get_permissions(self):
-        return self._permissions
-
-
-class JsonToken(wsme.types.Base):
-
-    """A token granting the bearer a limited set of permissions.
-
-    In all cases except creating a new token, the ``token`` attribute is empty.
-    There is no way to recover a lost token string except for revoking and
-    re-issuing the token.
-    """
-
-    _name = 'Token'
-
-    #: token ID
-    id = wsme.types.wsattr(int, mandatory=False)
-
-    #: the opaque token string (only set on new tokens)
-    token = wsme.types.wsattr(unicode, mandatory=False)
-
-    #: the user-supplied token description
-    description = wsme.types.wsattr(unicode, mandatory=True)
-
-    #: list of permissions this token grants
-    permissions = wsme.types.wsattr([unicode], mandatory=True)
-
-
 @bp.route('/')
 @login_required
 def root():
@@ -114,17 +50,17 @@ def root():
 
 @bp.route('/tokens')
 @p.base.tokens.view.require()
-@apimethod([JsonToken])
+@apimethod([types.JsonToken])
 def list_tokens():
     """Get a list of all existing tokens.
 
     Note that the response does not include the actual token strings.
     Such strings are only revealed when creating a new token."""
-    return [t.to_jsontoken() for t in Token.query.all()]
+    return [t.to_jsontoken() for t in tables.Token.query.all()]
 
 
 @bp.route('/tokens', methods=['POST'])
-@apimethod(JsonToken, body=JsonToken)
+@apimethod(types.JsonToken, body=types.JsonToken)
 @p.base.tokens.issue.require()
 def issue_token(body):
     """Issue a new token.  The body should not include a ``token`` or ``id``.
@@ -136,7 +72,7 @@ def issue_token(body):
         raise BadRequest("bad permissions")
 
     session = g.db.session('relengapi')
-    token_row = Token(
+    token_row = tables.Token(
         description=body.description,
         permissions=requested_permissions)
     session.add(token_row)
@@ -149,11 +85,11 @@ def issue_token(body):
 
 
 @bp.route('/tokens/<int:token_id>')
-@apimethod(JsonToken, int)
+@apimethod(types.JsonToken, int)
 @p.base.tokens.view.require()
 def get_token(token_id):
     """Get a token, identified by its ``id``."""
-    token_data = Token.query.filter_by(id=token_id).first()
+    token_data = tables.Token.query.filter_by(id=token_id).first()
     if not token_data:
         raise NotFound
     return token_data.to_jsontoken()
@@ -161,7 +97,7 @@ def get_token(token_id):
 
 @bp.route('/tokens/query', methods=['POST'])
 @p.base.tokens.view.require()
-@apimethod(JsonToken, body=unicode)
+@apimethod(types.JsonToken, body=unicode)
 def get_token_by_token(body):
     """Get a token, specified by the token key given in the request body
     (this avoids embedding a token in a URL, where it might be logged)."""
@@ -170,7 +106,7 @@ def get_token_by_token(body):
     if not claims:
         raise NotFound
 
-    token_data = Token.query.filter_by(id=claims['id']).first()
+    token_data = tables.Token.query.filter_by(id=claims['id']).first()
     if not token_data:
         raise NotFound
     return token_data.to_jsontoken()
@@ -182,30 +118,13 @@ def get_token_by_token(body):
 def revoke_token(token_id):
     """Revoke an authentication token, identified by its ID."""
     session = g.db.session('relengapi')
-    Token.query.filter_by(id=token_id).delete()
+    tables.Token.query.filter_by(id=token_id).delete()
     session.commit()
     return None, 204
 
 
-@auth.request_loader
-def token_loader(request):
-    # extract the token from the headers, returning None if anything's
-    # wrong
-    header = request.headers.get('Authentication')
-    if not header:
-        return
-    header = header.split()
-    if len(header) != 2 or header[0].lower() != 'bearer':
-        return
-    claims = tokenstr.str_to_claims(header[1])
-    if not claims:
-        return
-
-    token_data = Token.query.filter_by(id=claims['id']).first()
-    if token_data:
-        user = TokenUser(token_data.id, token_data.permissions)
-        logger.debug("Token access by %s", user)
-        return user
+# enable the loader to get a look at each incoming request
+auth.request_loader(loader.token_loader)
 
 
 @bp.record
